% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.backLinksClass.R
\name{print.backLinksClass}
\alias{print.backLinksClass}
\title{S3method print backLinks}
\usage{
\method{print}{backLinksClass}(x, maxprint = 10, ...)
}
\arguments{
\item{x}{the backLinksClass object, result of the backLinks function}

\item{maxprint}{the maximal number of pages to print}

\item{...}{others parameters}
}
\description{
S3method to print backLinksClass objects, results of the backLinks function
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}
\seealso{
backLinks
}

