% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.contribsClass.R
\name{print.contribsClass}
\alias{print.contribsClass}
\title{S3method print contribs}
\usage{
\method{print}{contribsClass}(x, maxprint = 10, ...)
}
\arguments{
\item{x}{the contribsClass object, result of the contribs function}

\item{maxprint}{the maximal number of pages to print}

\item{...}{others parameters}
}
\description{
S3method to print contribsClass objects, results of the contribs function
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}
\seealso{
contribs
}

