% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linksClass.R
\name{print.linksClass}
\alias{print.linksClass}
\title{S3method print links}
\usage{
\method{print}{linksClass}(x, maxprint = 10, ...)
}
\arguments{
\item{x}{the linksClass object, result of the links function}

\item{maxprint}{the maximal number of pages to print}

\item{...}{others parameters}
}
\description{
S3method to print linksClass objects, results of the links function
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}
\seealso{
links
}

