% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.userContribsClass.R
\name{print.userContribsClass}
\alias{print.userContribsClass}
\title{S3method print userContribs}
\usage{
\method{print}{userContribsClass}(x, maxprint = 10, ...)
}
\arguments{
\item{x}{the userContribsClass object, result of the userContribs function}

\item{maxprint}{the maximal number of contributions to print}

\item{...}{others parameters}
}
\description{
S3method to print userContribsClass objects, results of the userContribs function
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}
\seealso{
userContribs
}

