% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.userInfoClass.R
\name{print.userInfoClass}
\alias{print.userInfoClass}
\title{S3method print userInfo}
\usage{
\method{print}{userInfoClass}(x, ...)
}
\arguments{
\item{x}{the userInfoClass object, result of the userInfo function}

\item{...}{others parameters}
}
\description{
S3method to print userInfoClass objects, results of the userInfo function
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}
\seealso{
userInfo
}

