% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWikiPage.R
\name{testWikiPage}
\alias{testWikiPage}
\title{internal function testWikiPage}
\usage{
testWikiPage(page = NULL, domain = "en")
}
\arguments{
\item{page}{numeric identifier or character title of the specific wikipedia page}

\item{domain}{a character value providing the language of the wikipedia page.The default value is "en" for "english language".}
}
\value{
a list of four elements, the first is \code{takeOnlyFirst}, a boolean indicating if the class of \code{page} parameter is invalid, 
for example vector, list, matrix..., and in that case, the only the first element is considered.
 The second element is \code{redirPage}, the title page of the redirected page.
 The third element is \code{test}, an integer with value: 
\itemize{
\item{4}{ for invalid domain,}
\item{3}{ for an empty parameter page,}
\item{2}{ when Wikipedia does not have an article with this exact name,}
\item{1}{ for ambiguous page, direct or redirect,} 
\item{0}{ for valid an unambiguous page, direct or redirect. }
}
The last element, \code{warnMessage}, is a vector of warning messages.
}
\description{
internal function to test validity of a wikidepia page, called in \code{links}, \code{backLinks}, and \code{contribs} functions.
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}

