% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWikiUser.R
\name{testWikiUser}
\alias{testWikiUser}
\title{internal function testWikiUser}
\usage{
testWikiUser(user.name = NULL, domain = "en")
}
\arguments{
\item{user.name}{a character value providing the name of the specific wikipedia user}

\item{domain}{a character value providing the language of the wikipedia page.The default value is "en" for "english language".}
}
\value{
a list of four elements, the first is \code{takeOnlyFirst}, a boolean indicating if the class of \code{user.name} parameter is invalid, 
for example vector, list, matrix..., and in that case, the only the first element is considered.
 The second element is \code{test}, an integer with value: 
\itemize{
\item{4}{ for invalid domain,}
\item{3}{ for an empty parameter user.name,}
\item{2}{ when Wikipedia does not have an user with this exact name,}
\item{0}{ for valid existing user.name. }
}
The third element is \code{user}, a vector containing the user name, the user identifier, and domain. If the user does not exist, this item is not created.
The last element, \code{warnMessage}, is a vector of warning messages.
}
\description{
internal function to test validity of a wikidepia domain and user, 
called in \code{userInfo} and \code{userContribs} functions.
}
\author{
Avner Bar-Hen, Louise Baschet, Francois-Xavier Jollois, Jeremie Riou
}

